Feature/OT111-34 Create Endpoint Update Usuario. (#33)

+++ src/main/java/com/alkemy/ong/repository/RoleRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoleRepository extends JpaRepository<RoleEntity, Long> {
+public interface RoleRepository extends JpaRepository<RoleEntity, Integer> {
     RoleEntity findByName(String roleName);
 }
